Add SimpleTree
It's ordinary binary tree. NOT Search binary tree. I wrote function that search LCA of two given nodes in the tree. The lowest common ancestor is defined between two nodes p and q as the lowest node in T that has both p and q as descendants (where we allow a node to be a descendant of itself).